chore(package): Move player model into model.player package
