+++ src/main/java/edu/ntnu/idi/idatt/factory/view/PlayerTokenFactory.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idi.idatt.factory.view;
 
-import edu.ntnu.idi.idatt.model.PlayerTokenType;
+import edu.ntnu.idi.idatt.model.player.PlayerTokenType;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
 import javafx.scene.shape.Polygon;
