+++ src/main/java/edu/ntnu/idi/idatt/model/game/LadderBoardGame.java
@@ -3,7 +3,7 @@ package edu.ntnu.idi.idatt.model.game;
 import java.util.List;
 
 import edu.ntnu.idi.idatt.model.board.Board;
-import edu.ntnu.idi.idatt.model.Player;
+import edu.ntnu.idi.idatt.model.player.Player;
 import edu.ntnu.idi.idatt.model.tile.Tile;
 
 /**
