+++ src/main/java/edu/ntnu/idi/idatt/model/tile/LadderAction.java
@@ -1,7 +1,7 @@
 package edu.ntnu.idi.idatt.model.tile;
 
 import edu.ntnu.idi.idatt.model.board.Board;
-import edu.ntnu.idi.idatt.model.Player;
+import edu.ntnu.idi.idatt.model.player.Player;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionPerformValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDescriptionValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDestinationTileIdValidator;
