+++ src/main/java/edu/ntnu/idi/idatt/navigation/ViewNavigator.java
@@ -9,7 +9,7 @@ import edu.ntnu.idi.idatt.controller.BoardCreatorController;
 import edu.ntnu.idi.idatt.controller.LadderGameController;
 import edu.ntnu.idi.idatt.controller.MainMenuController;
 import edu.ntnu.idi.idatt.model.board.Board;
-import edu.ntnu.idi.idatt.model.Player;
+import edu.ntnu.idi.idatt.model.player.Player;
 import edu.ntnu.idi.idatt.observer.ButtonClickObserver;
 import edu.ntnu.idi.idatt.view.container.AppView;
 import edu.ntnu.idi.idatt.view.container.BoardCreatorView;
