+++ src/main/java/edu/ntnu/idi/idatt/observer/BoardGameSubject.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idi.idatt.observer;
 
-import edu.ntnu.idi.idatt.model.Player;
+import edu.ntnu.idi.idatt.model.player.Player;
 import edu.ntnu.idi.idatt.model.tile.TileAction;
 
 /**
