+++ src/main/java/edu/ntnu/idi/idatt/view/component/GameBoardStackPane.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 import edu.ntnu.idi.idatt.factory.view.PlayerTokenFactory;
 import edu.ntnu.idi.idatt.model.board.Board;
-import edu.ntnu.idi.idatt.model.Player;
+import edu.ntnu.idi.idatt.model.player.Player;
 import edu.ntnu.idi.idatt.model.tile.Tile;
 import edu.ntnu.idi.idatt.view.util.ViewUtils;
 import javafx.animation.PathTransition;
