+++ src/main/java/edu/ntnu/idi/idatt/view/container/LadderGameView.java
@@ -1,7 +1,7 @@
 package edu.ntnu.idi.idatt.view.container;
 
 import edu.ntnu.idi.idatt.model.board.Board;
-import edu.ntnu.idi.idatt.model.Player;
+import edu.ntnu.idi.idatt.model.player.Player;
 import edu.ntnu.idi.idatt.observer.ButtonClickObserver;
 import edu.ntnu.idi.idatt.observer.ButtonClickSubject;
 import edu.ntnu.idi.idatt.view.component.GameBoardStackPane;
